+++ webview-ui/src/components/mcp/configuration/McpView.tsx
