+++ webview-ui/src/components/mcp/configuration/tabs/installed/server-row/ServerRow.tsx
